+++ pkg/cmd/kind/create/cluster/validation/awsvalidator.go
@@ -38,8 +38,12 @@ func newAWSValidator(managed bool) *AWSValidator {
 	return awsInstance
 }
 
-func (v *AWSValidator) DescriptorFile(descriptorFile commons.DescriptorFile) {
-	v.descriptor = descriptorFile
+func (v *AWSValidator) Spec(spec commons.Spec) {
+	v.descriptor = spec
+}
+
+func (v *AWSValidator) DescriptorFile(spec commons.Spec) {
+	v.descriptor = spec
 }
 
 func (v *AWSValidator) SecretsFile(secrets commons.SecretsFile) {
@@ -72,22 +76,22 @@ func (v *AWSValidator) CommonsValidations() error {
 	return nil
 }
 
-func (v *AWSValidator) descriptorAwsValidations(descriptorFile commons.DescriptorFile, secretsFile commons.SecretsFile) error {
-	err := commonsDescriptorValidation(descriptorFile)
+func (v *AWSValidator) descriptorAwsValidations(spec commons.Spec, secretsFile commons.SecretsFile) error {
+	err := commonsDescriptorValidation(spec)
 	if err != nil {
 		return err
 	}
-	err = validateVPCCidr(descriptorFile)
+	err = validateVPCCidr(spec)
 	if err != nil {
 		return err
 	}
-	if descriptorFile.ControlPlane.Managed {
-		err = eksAZValidation(descriptorFile, secretsFile)
+	if spec.ControlPlane.Managed {
+		err = eksAZValidation(spec, secretsFile)
 		if err != nil {
 			return err
 		}
 	}
-	err = v.storageClassValidation(descriptorFile)
+	err = v.storageClassValidation(spec)
 	if err != nil {
 		return err
 	}
@@ -102,12 +106,12 @@ func secretsAwsValidations(secretsFile commons.SecretsFile) error {
 	return nil
 }
 
-func validateVPCCidr(descriptorFile commons.DescriptorFile) error {
-	if descriptorFile.Networks.PodsCidrBlock != "" {
+func validateVPCCidr(spec commons.Spec) error {
+	if spec.Networks.PodsCidrBlock != "" {
 		_, validRange1, _ := net.ParseCIDR("100.64.0.0/10")
 		_, validRange2, _ := net.ParseCIDR("198.19.0.0/16")
 
-		_, ipv4Net, _ := net.ParseCIDR(descriptorFile.Networks.PodsCidrBlock)
+		_, ipv4Net, _ := net.ParseCIDR(spec.Networks.PodsCidrBlock)
 
 		cidrSize := cidr.AddressCount(ipv4Net)
 		if cidrSize > cidrSizeMax || cidrSize < cidrSizeMin {
@@ -122,9 +126,9 @@ func validateVPCCidr(descriptorFile commons.DescriptorFile) error {
 	return nil
 }
 
-func eksAZValidation(descriptorFile commons.DescriptorFile, secretsFile commons.SecretsFile) error {
+func eksAZValidation(spec commons.Spec, secretsFile commons.SecretsFile) error {
 	awsCredentials := []string{
-		"AWS_REGION=" + descriptorFile.Region,
+		"AWS_REGION=" + spec.Region,
 		"AWS_ACCESS_KEY_ID=" + secretsFile.Secrets.AWS.Credentials.AccessKey,
 		"AWS_SECRET_ACCESS_KEY=" + secretsFile.Secrets.AWS.Credentials.SecretKey,
 	}
@@ -140,9 +144,9 @@ func eksAZValidation(descriptorFile commons.DescriptorFile, secretsFile commons.
 		return err
 	}
 	svc := ec2.New(sess)
-	if descriptorFile.Networks.Subnets != nil {
+	if spec.Networks.Subnets != nil {
 		privateAZs := []string{}
-		for _, subnet := range descriptorFile.Networks.Subnets {
+		for _, subnet := range spec.Networks.Subnets {
 			privateSubnetID, _ := filterPrivateSubnet(svc, &subnet.SubnetId)
 			if len(privateSubnetID) > 0 {
 				sid := &ec2.DescribeSubnetsInput{
@@ -160,9 +164,9 @@ func eksAZValidation(descriptorFile commons.DescriptorFile, secretsFile commons.
 			}
 		}
 		if len(privateAZs) < 3 {
-			return errors.New("Insufficient Availability Zones in region " + descriptorFile.Region + ". Please add at least 3 private subnets in different Availability Zones")
+			return errors.New("Insufficient Availability Zones in region " + spec.Region + ". Please add at least 3 private subnets in different Availability Zones")
 		}
-		for _, node := range descriptorFile.WorkerNodes {
+		for _, node := range spec.WorkerNodes {
 			if node.ZoneDistribution == "unbalanced" && node.AZ != "" {
 				if !slices.Contains(privateAZs, node.AZ) {
 					return errors.New("Worker node " + node.Name + " whose AZ is defined in " + node.AZ + " must match with the AZs associated to the defined subnets in descriptor")
@@ -175,7 +179,7 @@ func eksAZValidation(descriptorFile commons.DescriptorFile, secretsFile commons.
 			return err
 		}
 		if len(result.AvailabilityZones) < 3 {
-			return errors.New("Insufficient Availability Zones in region " + descriptorFile.Region + ". Must have at least 3")
+			return errors.New("Insufficient Availability Zones in region " + spec.Region + ". Must have at least 3")
 		}
 		azs := make([]string, 3)
 		for i, az := range result.AvailabilityZones {
@@ -184,10 +188,10 @@ func eksAZValidation(descriptorFile commons.DescriptorFile, secretsFile commons.
 			}
 			azs[i] = *az.ZoneName
 		}
-		for _, node := range descriptorFile.WorkerNodes {
+		for _, node := range spec.WorkerNodes {
 			if node.ZoneDistribution == "unbalanced" && node.AZ != "" {
 				if !slices.Contains(azs, node.AZ) {
-					return errors.New("Worker node " + node.Name + " whose AZ is defined in " + node.AZ + " must match with the first three AZs in region " + descriptorFile.Region)
+					return errors.New("Worker node " + node.Name + " whose AZ is defined in " + node.AZ + " must match with the first three AZs in region " + spec.Region)
 				}
 			}
 		}
@@ -228,14 +232,14 @@ func filterPrivateSubnet(svc *ec2.EC2, subnetID *string) (string, error) {
 	}
 }
 
-func (v *AWSValidator) storageClassValidation(descriptorFile commons.DescriptorFile) error {
-	if descriptorFile.StorageClass.EncryptionKey != "" {
-		err := v.storageClassKeyFormatValidation(descriptorFile.StorageClass.EncryptionKey)
+func (v *AWSValidator) storageClassValidation(spec commons.Spec) error {
+	if spec.StorageClass.EncryptionKey != "" {
+		err := v.storageClassKeyFormatValidation(spec.StorageClass.EncryptionKey)
 		if err != nil {
 			return errors.New("Error in StorageClass: " + err.Error())
 		}
 	}
-	err := v.storageClassParametersValidation(descriptorFile)
+	err := v.storageClassParametersValidation(spec)
 	if err != nil {
 		return errors.New("Error in StorageClass: " + err.Error())
 	}
@@ -251,11 +255,11 @@ func (v *AWSValidator) storageClassKeyFormatValidation(key string) error {
 	return nil
 }
 
-func (v *AWSValidator) storageClassParametersValidation(descriptorFile commons.DescriptorFile) error {
-	sc := descriptorFile.StorageClass
+func (v *AWSValidator) storageClassParametersValidation(spec commons.Spec) error {
+	sc := spec.StorageClass
 	typesSupportedForIOPS := []string{"io1", "io2", "gp3"}
 	fstypes := []string{"xfs", "ext3", "ext4", "ext2"}
-	err := verifyFields(descriptorFile)
+	err := verifyFields(spec)
 	if err != nil {
 		return err
 	}